html/browsers/windows/embedded-opener-remove-frame.html

<!doctype html>
<title>opener and discarded browsing contexts</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<iframe name=matchesastring></iframe>
<script>
function testOpener(otherW, thisW, discardOtherBC) {
  assert_equals(otherW.opener, thisW, "opener before removal");

  const openerDesc = Object.getOwnPropertyDescriptor(otherW, "opener"),
        openerGet = openerDesc.get;

  assert_equals(openerGet(), thisW, "opener before removal via directly invoking the getter");
  discardOtherBC();
  assert_equals(otherW.opener, null, "opener after removal");
  assert_equals(openerGet(), null, "opener after removal via directly invoking the getter");

  otherW.opener = null;
  assert_equals(openerGet(), null, "opener after setting it null via directly invoking the getter");
  const openerDescNull = Object.getOwnPropertyDescriptor(otherW, "opener");
  assert_not_equals(openerDescNull, openerDesc);
  assert_object_equals(openerDescNull, openerDesc);

  otherW.opener = "immaterial";
  assert_equals(openerGet(), null, "opener after setting it \"immaterial\" via directly invoking the getter");
  const openerDescImmaterial = Object.getOwnPropertyDescriptor(otherW, "opener");
  assert_equals(openerDescImmaterial.value, "immaterial");
  assert_true(openerDescImmaterial.writable);
  assert_true(openerDescImmaterial.enumerable);
  assert_true(openerDescImmaterial.configurable);
}

async_test(t => {
  const frame = document.querySelector("iframe"),
        frameW = frame.contentWindow;
  frame.onload = t.step_func(() => {
    // Firefox and Chrome/Safari load differently
    if (frame.contentWindow.location.href === "about:blank") {
      return;
    }

    // Test bits
    testOpener(frameW, window, () => frame.remove());

    t.done();
  });
  window.open("/common/blank.html", "matchesastring");
}, "opener of discarded nested browsing context");

async_test(t => {
  const popupW = window.open("/common/blank.html");
  popupW.onload = t.step_func_done(t => {
    testOpener(popupW, window, () => popupW.close());
  });
}, "opener of discarded auxiliary browsing context");
</script>
