third_party/blink/web_tests/external/wpt/html/browsers/windows/embedded-opener-remove-frame-expected.txt

This is a testharness.js-based test.
[FAIL] opener of discarded nested browsing context
  assert_equals: expected (string) "immaterial" but got (undefined) undefined
[FAIL] opener of discarded auxiliary browsing context
  assert_equals: expected (string) "immaterial" but got (undefined) undefined
Harness: the test ran to completion.
